requirements.txt
allure-python-commons
playwright>=1.47
pytest>=8.2
pytest-playwright>=0.6
pytest-xdist>=3.5
pytest-base-url>=2.0

allure-pytest>=2.13

python-dotenv>=1.0
faker>=25.0

ruff>=0.6
black>=24.0
isort>=5.13